authservice-readable/requirements.txt
Flask~=2.2.3
flask-redis

requests~=2.28.2
PyJWT~=2.6.0